trading_system/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0
alpaca-py>=0.28.0
yfinance>=0.2.40
ta>=0.11.0
transformers>=4.40.0
torch>=2.2.0
vectorbt>=0.26.0
pandas>=2.2.0
numpy>=1.26.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
# MCP & Utils
httpx>=0.27.0
sse-starlette>=2.0.0
duckduckgo-search>=6.0.0
supabase>=2.0.0
slack-sdk>=3.27.0
pytest-asyncio>=0.23.0
